requirements.txt
pandas
numpy
scikit-learn
matplotlib
seaborn
pyyaml
imbalanced-learn
requests
hydra-core==1.3.2
omegaconf==2.3.0
mlflow==2.20.3
fastapi==0.104.0
uvicorn[standard]==0.24.0
wandb
python-dotenv
mlflow-skinny 